backend/vendor_connect/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
atpublic==4.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
Django==5.0
django-filter==23.5
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
frozenlist==1.4.1
GDAL==3.6.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
jwcrypto==1.5.1
Markdown==3.5.1
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
phonenumbers==8.13.27
Pillow==10.1.0
promise==2.3
psycopg==3.1.16
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
twilio==8.11.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
world==5.1.1
wrapt==1.16.0
yarl==1.9.4
google-auth